move functions and cleanup unused args
